Move code from pages component to index file
